add func 'payload.NewFile' which can be used to create Payload from a local file.
