+++ payload/payload_test.go
@@ -1,8 +1,10 @@
 package payload
 
 import (
-	"github.com/stretchr/testify/assert"
+	"fmt"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestPayload_new(t *testing.T) {
@@ -18,4 +20,10 @@ func TestPayload_new(t *testing.T) {
 	metadata2, ok = p1.MetadataUTF8()
 	assert.False(t, ok)
 	assert.Equal(t, "", metadata2)
+}
+
+func TestNewFile(t *testing.T) {
+	pl, err := NewFile("/etc/hosts", nil)
+	assert.NoError(t, err, "bad file")
+	fmt.Print(pl.DataUTF8())
 }
